# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Visa requirement checker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Country {
  /** ISO 3166-1 alpha-2 country code */
  code: string;
  /** Country name */
  name: string;
  /** Unicode flag emoji */
  flag: string;
  /** Geographic region */
  region: string;
}

/**
 * The type of visa requirement
 */
export type VisaRequirement =
  (typeof VisaRequirement)[keyof typeof VisaRequirement];

export const VisaRequirement = {
  visa_free: "visa_free",
  visa_on_arrival: "visa_on_arrival",
  e_visa: "e_visa",
  visa_required: "visa_required",
  no_admission: "no_admission",
} as const;

export interface VisaResult {
  passportCountry: Country;
  destinationCountry: Country;
  requirement: VisaRequirement;
  /** Maximum allowed stay (e.g. "90 days") */
  maxStay?: string | null;
  /** Additional notes about entry requirements */
  notes?: string | null;
}

export interface CheckVisaMultipleBody {
  /** ISO 3166-1 alpha-2 passport country code */
  passport: string;
  /** List of ISO 3166-1 alpha-2 destination country codes */
  destinations: string[];
}

export interface VisaStats {
  passportCountry: Country;
  /** Number of countries accessible visa-free */
  visaFree: number;
  /** Number of countries accessible with visa on arrival */
  visaOnArrival: number;
  /** Number of countries accessible with e-visa */
  eVisa: number;
  /** Number of countries requiring a visa */
  visaRequired: number;
  /** Number of countries with no admission */
  noAdmission: number;
  /** Total countries in the index */
  total: number;
  /** Passport power ranking (1 = strongest) */
  powerRank: number;
}

export interface VisaDetail {
  /** Visa fee in US dollars (0 for free, null for not applicable) */
  feeUSD: number | null;
  /** How long processing takes */
  processingDays: string;
  /** Maximum allowed stay */
  maxStay: string;
  /** List of required documents */
  documents: string[];
  /** Step-by-step application process */
  process: string[];
  /** URL to apply for visa online */
  onlineApplyUrl?: string | null;
  /** Additional notes */
  notes?: string | null;
}

/**
 * Type of attraction
 */
export type AttractionType =
  (typeof AttractionType)[keyof typeof AttractionType];

export const AttractionType = {
  landmark: "landmark",
  nature: "nature",
  museum: "museum",
  beach: "beach",
  city: "city",
  heritage: "heritage",
  temple: "temple",
} as const;

export interface Attraction {
  /** Name of the attraction */
  name: string;
  /** Brief description of the attraction */
  description: string;
  /** Wikipedia article title for fetching images */
  wikiTitle: string;
  /** Type of attraction */
  type: AttractionType;
  /** Suggested time to spend at this attraction */
  visitDuration: string;
}

export interface CountryTouristInfo {
  tagline: string;
  bestTimeToVisit: string;
  currency: string;
  language: string;
  capital: string;
  timezone: string;
  attractions: Attraction[];
}

export interface OfficialLinks {
  /** Official government visa application portal URL for the destination country */
  visaPortal: string;
  /** Official MFA/embassy finder URL listing the destination country's missions abroad */
  embassyFinder: string;
}

export interface DestinationInfo {
  passportCountry: Country;
  destinationCountry: Country;
  requirement: VisaRequirement;
  maxStay?: string | null;
  notes?: string | null;
  visaDetail: VisaDetail;
  touristInfo?: CountryTouristInfo | null;
  officialLinks?: OfficialLinks | null;
}

export interface ErrorResponse {
  error: string;
}

export type CheckVisaParams = {
  /**
   * ISO 3166-1 alpha-2 passport country code (e.g. "US")
   */
  passport: string;
  /**
   * ISO 3166-1 alpha-2 destination country code (e.g. "FR")
   */
  destination: string;
};

export type CheckVisaAllParams = {
  /**
   * ISO 3166-1 alpha-2 passport country code
   */
  passport: string;
};

export type GetDestinationInfoParams = {
  /**
   * ISO 3166-1 alpha-2 passport country code
   */
  passport: string;
  /**
   * ISO 3166-1 alpha-2 destination country code
   */
  destination: string;
};

export type GetVisaStatsParams = {
  /**
   * ISO 3166-1 alpha-2 passport country code
   */
  passport: string;
};