# lib/api-zod/src/generated/types/visaRequirement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Visa requirement checker API
 * OpenAPI spec version: 0.1.0
 */

/**
 * The type of visa requirement
 */
export type VisaRequirement =
  (typeof VisaRequirement)[keyof typeof VisaRequirement];

export const VisaRequirement = {
  visa_free: "visa_free",
  visa_on_arrival: "visa_on_arrival",
  e_visa: "e_visa",
  visa_required: "visa_required",
  no_admission: "no_admission",
} as const;